use PresetSelect in Menu

+++ inc/gui.hpp
@@ -157,7 +157,8 @@ public:
 	 * @pre presets was managed by PresetSelect
 	 */
 	PresetSelect(const App::Presets<TValue>& presets, size_t active) :
-	    presets{presets}, valueBuffer{presets.get(active).value}, textList{Range{presets.get()}}, activeItem(active),
+	    presets{presets}, valueBuffer{presets.get(active).value}, textList{Range{presets.get()}},
+	    activeItem(static_cast<int>(active)),
 	    state{presets.size() != 0 && presets.find(customName) == presets.size() - 1 ? State::Custom : State::Normal}
 	{
 	}
